+++ spec/internal/db/migrate/20260326234116_create_ductwork_advancements.rb
@@ -7,7 +7,7 @@ class CreateDuctworkAdvancements < Ductwork::Migration
         table,
         :process,
         index: true,
-        null: false,
+        null: true,
         foreign_key: { to_table: :ductwork_processes }
       )
       belongs_to(